163405 Intial changes for issue 119/121
Added new PS file which reads amiID from file path determined by Image Type
and for specific AWS regions and update the template file which will be
accessed using Template URL.
If Template URL is not specified, script will exit with an error.
Still I need to update the script with template update logic which will
be in next commit.